change price to integer

+++ db/migrate/20170401211841_create_articles.rb
@@ -3,7 +3,7 @@ class CreateArticles < ActiveRecord::Migration
     create_table :articles do |t|
       t.string :title, null: false
       t.string :description, null: false
-      t.money :price, null: false
+      t.integer :price, null: false
       t.string :email, null: false
       t.integer :category_id, null: false
       t.string :secret_key, null: false
